+++ cvodes/source/cvsspgmr.c
@@ -516,7 +516,7 @@ int CVSpgmrGetRealWorkSpace(void *cvode_mem, long int *lenrwSG)
 
 /*************** CVSpgmrGetNumPrecEvals ******************************/
 
-int CVSpgmrGetNumPrecEvals(void *cvode_mem, int *npevals)
+int CVSpgmrGetNumPrecEvals(void *cvode_mem, long int *npevals)
 {
   CVodeMem cv_mem;
   CVSpgmrMem cvspgmr_mem;
@@ -541,7 +541,7 @@ int CVSpgmrGetNumPrecEvals(void *cvode_mem, int *npevals)
 
 /*************** CVSpgmrGetNumPrecSolves *****************************/
 
-int CVSpgmrGetNumPrecSolves(void *cvode_mem, int *npsolves)
+int CVSpgmrGetNumPrecSolves(void *cvode_mem, long int *npsolves)
 {
   CVodeMem cv_mem;
   CVSpgmrMem cvspgmr_mem;
@@ -566,7 +566,7 @@ int CVSpgmrGetNumPrecSolves(void *cvode_mem, int *npsolves)
 
 /*************** CVSpgmrGetNumLinIters *******************************/
 
-int CVSpgmrGetNumLinIters(void *cvode_mem, int *nliters)
+int CVSpgmrGetNumLinIters(void *cvode_mem, long int *nliters)
 {
   CVodeMem cv_mem;
   CVSpgmrMem cvspgmr_mem;
@@ -591,7 +591,7 @@ int CVSpgmrGetNumLinIters(void *cvode_mem, int *nliters)
 
 /*************** CVSpgmrGetNumConvFails ******************************/
 
-int CVSpgmrGetNumConvFails(void *cvode_mem, int *nlcfails)
+int CVSpgmrGetNumConvFails(void *cvode_mem, long int *nlcfails)
 {
   CVodeMem cv_mem;
   CVSpgmrMem cvspgmr_mem;
@@ -616,7 +616,7 @@ int CVSpgmrGetNumConvFails(void *cvode_mem, int *nlcfails)
 
 /*************** CVSpgmrGetNumJtimesEvals ****************************/
 
-int CVSpgmrGetNumJtimesEvals(void *cvode_mem, int *njvevals)
+int CVSpgmrGetNumJtimesEvals(void *cvode_mem, long int *njvevals)
 {
   CVodeMem cv_mem;
   CVSpgmrMem cvspgmr_mem;
@@ -641,7 +641,7 @@ int CVSpgmrGetNumJtimesEvals(void *cvode_mem, int *njvevals)
 
 /*************** CVSpgmrGetNumRhsEvals *******************************/
 
-int CVSpgmrGetNumRhsEvals(void *cvode_mem, int *nfevalsSG)
+int CVSpgmrGetNumRhsEvals(void *cvode_mem, long int *nfevalsSG)
 {
   CVodeMem cv_mem;
   CVSpgmrMem cvspgmr_mem;
